requirements.txt
Faker==24.8.0
pandas==2.2.1
python-dateutil==2.9.0.post0
black==24.1.0
isort==5.12.0
flake8==6.1.0